vk-1502-rmvc-refactor: split RMVC into a bunch of controllers, wired them into NavigationViewController, removed RouteMapViewController, moved methods to NavigationMapVIew and separate extensions where possible.
