+++ Sources/MapboxNavigation/MapView.swift
@@ -28,6 +28,28 @@ extension MapView {
     }
     
     /**
+     Method, which returns identifiers of the tile sets that make up specific source.
+     
+     This array contains multiple entries for a composited source. This property is empty for non-Mapbox-hosted tile sets and sources with type other than `vector`.
+     */
+    func tileSetIdentifiers(_ sourceIdentifier: String, sourceType: String) -> [String] {
+        do {
+            if sourceType == "vector",
+               let properties = try __map.getStyleSourceProperties(forSourceId: sourceIdentifier).value as? Dictionary<String, Any>,
+               let url = properties["url"] as? String,
+               let configurationURL = URL(string: url),
+               configurationURL.scheme == "mapbox",
+               let tileSetIdentifiers = configurationURL.host?.components(separatedBy: ",") {
+                return tileSetIdentifiers
+            }
+        } catch {
+            NSLog("Failed to get source properties with error: \(error.localizedDescription).")
+        }
+        
+        return []
+    }
+    
+    /**
      Returns a Boolean value indicating whether data from the given tile set layer is currently visible in the map view’s style.
      
      - parameter tileSetIdentifier: Identifier of the tile set in the form `user.tileset`.
@@ -137,5 +159,19 @@ extension MapView {
         }
         
         return []
+    }
+    
+    /**
+     Method, which returns list of source identifiers, which contain streets tile set.
+     */
+    func streetsSources() -> [StyleObjectInfo] {
+        let streetsSources = (try? __map.getStyleSources().compactMap {
+            $0
+        }.filter {
+            let identifiers = tileSetIdentifiers($0.id, sourceType: $0.type)
+            return VectorSource.isMapboxStreets(identifiers)
+        }) ?? []
+        
+        return streetsSources
     }
 }
